add svg for shield, change color of title, make site usable on mobile
So I ran into an issue for the SVG where I only wanted to use just one old FA4 icon (the shield), but to do so I had to install this entire directory. This is probably a bad way to do things but idk, lmk if it's an issue (it probably is)

Also, the site should be useable on smaller screens now.